bugs repaired

+++ src/read_civdata.m
@@ -120,8 +120,8 @@ if vardetect(1)==0
      return
 end
 if strcmp(Data.Conventions,'uvmat/civdata/compress')
-    Field.X=Field.X-0.5+Field.U/2;% shift to the convected position
-    Field.Y=Field.Y-0.5+Field.V/2;
+    Field.X=double(Field.X)-0.5+Field.U/2;% shift to the convected position
+    Field.Y=double(Field.Y)-0.5+Field.V/2;
 end
 switch VelTypeOut
     case {'civ1','filter1'}
